refactor(wagmi): refactor wagmiConfig
